+++ pybamm/input/parameters/lithium_ion/negative_electrodes/graphite_ORegan2022/graphite_LGM50_ocp_Chen2020.py
@@ -3,8 +3,8 @@ from pybamm import exp, tanh
 
 def graphite_LGM50_ocp_Chen2020(sto):
     """
-    LG M50 Graphite open-circuit potential as a function of stochiometry. The fit is
-    taken from [1].
+    LG M50 Graphite open-circuit potential as a function of stochiometry, fit taken
+    from [1].
 
     References
     ----------