+++ pybamm/models/full_battery_models/lithium_ion/newman_tobias.py
@@ -18,7 +18,8 @@ class NewmanTobias(DFN):
     Parameters
     ----------
     options : dict, optional
-        A dictionary of options to be passed to the model.
+        A dictionary of options to be passed to the model. For a detailed list of
+        options see :class:`~pybamm.BatteryModelOptions`.
     name : str, optional
         The name of the model.
     build :  bool, optional
@@ -61,8 +62,8 @@ class NewmanTobias(DFN):
             if particle == "Fickian diffusion":
                 self.submodels[
                     domain.lower() + " particle"
-                ] = pybamm.particle.no_distribution.XAveragedFickianDiffusion(
-                    self.param, domain, self.options
+                ] = pybamm.particle.FickianDiffusion(
+                    self.param, domain, self.options, x_average=True
                 )
             elif particle in [
                 "uniform profile",
@@ -71,8 +72,8 @@ class NewmanTobias(DFN):
             ]:
                 self.submodels[
                     domain.lower() + " particle"
-                ] = pybamm.particle.no_distribution.XAveragedPolynomialProfile(
-                    self.param, domain, particle, self.options
+                ] = pybamm.particle.XAveragedPolynomialProfile(
+                    self.param, domain, self.options
                 )
 
     def set_electrolyte_submodel(self):