+++ pybamm/models/submodels/interface/kinetics/diffusion_limited.py
@@ -29,6 +29,9 @@ class DiffusionLimited(BaseInterface):
         self.order = order
 
     def get_coupled_variables(self, variables):
+        Domain = self.domain
+        rxn = self.reaction_name
+
         delta_phi_s = variables[self.domain + " electrode surface potential difference"]
         # If delta_phi_s was broadcast, take only the orphan
         if isinstance(delta_phi_s, pybamm.Broadcast):
@@ -37,7 +40,7 @@ class DiffusionLimited(BaseInterface):
         # Get exchange-current density
         j0 = self._get_exchange_current_density(variables)
         # Get open-circuit potential variables and reaction overpotential
-        ocp, dUdT = self._get_open_circuit_potential(variables)
+        ocp = variables[f"{Domain} electrode{rxn} open circuit potential"]
         eta_r = delta_phi_s - ocp
 
         # Get interfacial current densities
@@ -50,26 +53,14 @@ class DiffusionLimited(BaseInterface):
         )
         variables.update(self._get_standard_exchange_current_variables(j0))
         variables.update(self._get_standard_overpotential_variables(eta_r))
-        variables.update(self._get_standard_ocp_variables(ocp, dUdT))
+
+        variables.update(
+            self._get_standard_volumetric_current_density_variables(variables)
+        )
 
         # No SEI film resistance in this model
         eta_sei = pybamm.Scalar(0)
         variables.update(self._get_standard_sei_film_overpotential_variables(eta_sei))
-
-        if (
-            "Negative electrode" + self.reaction_name + " interfacial current density"
-            in variables
-            and "Positive electrode"
-            + self.reaction_name
-            + " interfacial current density"
-            in variables
-        ):
-            variables.update(
-                self._get_standard_whole_cell_interfacial_current_variables(variables)
-            )
-            variables.update(
-                self._get_standard_whole_cell_exchange_current_variables(variables)
-            )
 
         if self.order == "composite":
             # For the composite model, adds the first-order x-averaged interfacial