+++ tests/unit/test_plotting/test_quick_plot.py
@@ -16,7 +16,7 @@ class TestQuickPlot(unittest.TestCase):
         c = pybamm.Variable("c", domain=[])
 
         # Simple ODEs
-        model.rhs = {a: pybamm.Scalar(2), b: pybamm.Scalar(0), c: -c}
+        model.rhs = {a: pybamm.Scalar(0.2), b: pybamm.Scalar(0), c: -c}
 
         # Simple initial conditions
         model.initial_conditions = {
@@ -28,6 +28,7 @@ class TestQuickPlot(unittest.TestCase):
         # Broadcast some of the variables
         model.variables = {
             "a": a,
+            "b": b,
             "b broadcasted": pybamm.FullBroadcast(b, whole_cell, "current collector"),
             "c broadcasted": pybamm.FullBroadcast(
                 c, ["negative electrode", "separator"], "current collector"
@@ -38,6 +39,11 @@ class TestQuickPlot(unittest.TestCase):
             "c broadcasted positive electrode": pybamm.PrimaryBroadcast(
                 c, "positive particle"
             ),
+            "Variable with a very long name": a,
+            "2D variable": pybamm.FullBroadcast(
+                1, "negative particle", {"secondary": "negative electrode"}
+            ),
+            "NaN variable": pybamm.Scalar(np.nan),
         }
         model._timescale = pybamm.Scalar(1)
 
@@ -115,7 +121,6 @@ class TestQuickPlot(unittest.TestCase):
         quick_plot.slider_update(0.01)
 
         # Test longer name
-        model.variables["Variable with a very long name"] = model.variables["a"]
         quick_plot = pybamm.QuickPlot(solution, ["Variable with a very long name"])
         quick_plot.plot(0)
 
@@ -146,7 +151,7 @@ class TestQuickPlot(unittest.TestCase):
             quick_plot.plots[("a",)][0][0].get_xdata(), t_eval
         )
         np.testing.assert_array_almost_equal(
-            quick_plot.plots[("a",)][0][0].get_ydata(), 2 * t_eval
+            quick_plot.plots[("a",)][0][0].get_ydata(), 0.2 * t_eval
         )
         quick_plot = pybamm.QuickPlot(solution, ["a"], time_unit="minutes")
         quick_plot.plot(0)
@@ -155,7 +160,7 @@ class TestQuickPlot(unittest.TestCase):
             quick_plot.plots[("a",)][0][0].get_xdata(), t_eval / 60
         )
         np.testing.assert_array_almost_equal(
-            quick_plot.plots[("a",)][0][0].get_ydata(), 2 * t_eval
+            quick_plot.plots[("a",)][0][0].get_ydata(), 0.2 * t_eval
         )
         quick_plot = pybamm.QuickPlot(solution, ["a"], time_unit="hours")
         quick_plot.plot(0)
@@ -164,7 +169,7 @@ class TestQuickPlot(unittest.TestCase):
             quick_plot.plots[("a",)][0][0].get_xdata(), t_eval / 3600
         )
         np.testing.assert_array_almost_equal(
-            quick_plot.plots[("a",)][0][0].get_ydata(), 2 * t_eval
+            quick_plot.plots[("a",)][0][0].get_ydata(), 0.2 * t_eval
         )
         with self.assertRaisesRegex(ValueError, "time unit"):
             pybamm.QuickPlot(solution, ["a"], time_unit="bad unit")
@@ -186,11 +191,6 @@ class TestQuickPlot(unittest.TestCase):
             pybamm.QuickPlot(solution, ["a"], spatial_unit="bad unit")
 
         # Test 2D variables
-        model.variables["2D variable"] = disc.process_symbol(
-            pybamm.FullBroadcast(
-                1, "negative particle", {"secondary": "negative electrode"}
-            )
-        )
         quick_plot = pybamm.QuickPlot(solution, ["2D variable"])
         quick_plot.plot(0)
         quick_plot.dynamic_plot(testing=True)
@@ -255,7 +255,6 @@ class TestQuickPlot(unittest.TestCase):
             )
 
         # No variable can be NaN
-        model.variables["NaN variable"] = disc.process_symbol(pybamm.Scalar(np.nan))
         with self.assertRaisesRegex(
             ValueError, "All-NaN variable 'NaN variable' provided"
         ):